fix: 홈 뷰모델 구조 변경
