+++ app/src/main/java/com/zenefit/zenefit_android/presentation/ui/main/home/HomeFragment.kt
@@ -8,19 +8,21 @@ import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import com.zenefit.zenefit_android.R
 import com.zenefit.zenefit_android.data.remote.response.ResponseHomeData
 import com.zenefit.zenefit_android.data.remote.response.ResponsePolicyCountData
 import com.zenefit.zenefit_android.databinding.FragmentHomeBinding
 import com.zenefit.zenefit_android.presentation.ui.main.home.adapter.HomePolicyDeadlineAdapter
 import com.zenefit.zenefit_android.presentation.ui.main.home.adapter.HomePolicyRecommendAdapter
+import com.zenefit.zenefit_android.presentation.ui.main.home.viewmodel.HomeViewModel
 import com.zenefit.zenefit_android.presentation.ui.main.viewmodel.MainViewModel
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class HomeFragment : Fragment() {
     private lateinit var binding : FragmentHomeBinding
-    private val viewModel : MainViewModel by activityViewModels()
+    private val viewModel : HomeViewModel by viewModels()
 
     private val recommendPolicyAdapter by lazy {
         HomePolicyRecommendAdapter(::onRecommendApplyClicked)
